+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `wxa_sdk check` command-line command to invoke a skill's health check
 - Added this changelog
 
+### Changed
+- Changed response serialization in server to be compatible with mindmeld 4.3.4
+
 ## 0.1.1 - 2020-02-19
 
 - Initial implementation